#248 normalize_pathをutilsに異動
