+++ Makefile
@@ -58,7 +58,8 @@ SRCS_UTILS = $(UTILS_DIR)/free.c \
 	$(UTILS_DIR)/list.c \
 	$(UTILS_DIR)/string.c \
 	$(UTILS_DIR)/gnl.c \
-	$(UTILS_DIR)/pwd_utils.c
+	$(UTILS_DIR)/pwd_utils.c \
+	$(UTILS_DIR)/normalize_path.c
 SRCS_BUILTINS = $(BUILTINS_DIR)/builtin_pwd.c \
 	$(BUILTINS_DIR)/builtin_echo.c \
 	$(BUILTINS_DIR)/builtin_cd.c \
