+++ includes/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aomatsud <aomatsud@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: hmaruyam <hmaruyam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/16 16:18:32 by aomatsud          #+#    #+#             */
-/*   Updated: 2025/10/16 18:59:47 by aomatsud         ###   ########.fr       */
+/*   Updated: 2025/10/29 00:01:35 by hmaruyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,8 +42,8 @@ void						print_error_msg_builtin(char *cmd, char *context,
 								t_blt_error error);
 int							error_lst(t_list *lst, char *context,
 								t_status status, void (*del)(void *));
-int							error_parent(t_pipeline *pipeline,
-								char *context, t_status status);
+int							error_parent(t_pipeline *pipeline, char *context,
+								t_status status);
 void						exit_error(t_minishell *minishell,
 								t_pipeline *pipeline, char *context,
 								t_status status);
@@ -51,5 +51,7 @@ void						exit_success(t_minishell *minishell,
 								t_pipeline *pipeline);
 
 int							is_pwd_valid(char *pwd_path);
+
+char						*normalize_path(const char *abs_path);
 
 #endif
